Refactor 'createTestActions'
The refactor breaks the fold up in to two stages:

  * Building a test tree with initializers, finalizers, and actions with
    meta information: test path, test dependencies, and test status.

  * Collecting initializers and finalizers in a single pass, and
    applying them to every test action.

This removes an accidental quadratic complexity in favor of a linear
one caused by continuously mapping over tests while building that
list of tests. This removes the need for one layer in `Tr`'s monad
stack: `WriterT`. At the same time, it uses dedicated data structures
instead of tuples for ease of reading.

This commit aims to make future extensions easier.